requirement.txt
paddlepaddle
visualdl
opencv-python
matplotlib
